r3147  - printSelect(): switch to a new spec to allow varying number of HTML attributes for the SELECT element  - printNiftySelect(): idem

+++ inc/interface.php
@@ -561,7 +561,7 @@ function renderEditObjectForm ($object_id)
 	echo '<table border=0 cellspacing=0 cellpadding=3 align=center>';
 	echo "<tr><td>&nbsp;</td><th colspan=2><h2>Attributes</h2></th></tr>";
 	echo "<tr><td>&nbsp;</td><th class=tdright>Type:</th><td class=tdleft>";
-	printNiftySelect (cookOptgroups (readChapter (CHAP_OBJTYPE, 'o')), 'object_type_id', $object['objtype_id']);
+	printNiftySelect (cookOptgroups (readChapter (CHAP_OBJTYPE, 'o')), array ('name' => 'object_type_id'), $object['objtype_id']);
 	echo "</td></tr>\n";
 	// baseline info
 	echo "<tr><td>&nbsp;</td><th class=tdright>Common name:</th><td class=tdleft><input type=text name=object_name value='${object['name']}'></td></tr>\n";
@@ -599,7 +599,7 @@ function renderEditObjectForm ($object_id)
 					$chapter = readChapter ($record['chapter_id'], 'o');
 					$chapter[0] = '-- NOT SET --';
 					$chapter = cookOptgroups ($chapter, $object['objtype_id'], $record['key']);
-					printNiftySelect ($chapter, "${i}_value", $record['key']);
+					printNiftySelect ($chapter, array ('name' => "${i}_value"), $record['key']);
 					break;
 			}
 			echo "</td></tr>\n";
@@ -638,7 +638,7 @@ function renderEditRackForm ($rack_id)
 	printOpFormIntro ('updateRack');
 	echo '<table border=0 align=center>';
 	echo "<tr><th class=tdright>Rack row:</th><td class=tdleft>";
-	printSelect (getRackRows(), 'rack_row_id', $rack['row_id']);
+	printSelect (getRackRows(), array ('name' => 'rack_row_id'), $rack['row_id']);
 	echo "</td></tr>\n";
 	echo "<tr><th class=tdright>Name (required):</th><td class=tdleft><input type=text name=rack_name value='${rack['name']}'></td></tr>\n";
 	echo "<tr><th class=tdright>Height (required):</th><td class=tdleft><input type=text name=rack_height value='${rack['height']}'></td></tr>\n";
@@ -662,33 +662,43 @@ function renderEditRackForm ($rack_id)
 
 // This is a helper for creators and editors.
 // Input array keys are OPTION VALUEs and input array values are OPTION text.
-function printSelect ($optionList, $select_name, $selected_id = NULL, $tabindex = NULL)
+function printSelect ($optionList, $select_attrs = array(), $selected_id = NULL)
 {
-	echo "<select name=${select_name}" . ($tabindex ? " tabindex=${tabindex}" : '') . '>';
+	if (!array_key_exists ($select_attrs, 'name'))
+		return;
+	echo '<select';
+	foreach ($select_attrs as $attr_name => $attr_value)
+		echo " ${attr_name}=${attr_value}";
+	echo '>';
 	foreach ($optionList as $dict_key => $dict_value)
 		echo "<option value='${dict_key}'" . ($dict_key == $selected_id ? ' selected' : '') . ">${dict_value}</option>";
-	echo "</select>";
+	echo '</select>';
 }
 
 // Input is a cooked list of OPTGROUPs, each with own sub-list of OPTIONs in the same
 // format as printSelect() expects.
-function printNiftySelect ($groupList, $select_name, $selected_id = NULL, $tabindex = NULL)
+function printNiftySelect ($groupList, $select_attrs, $selected_id = NULL)
 {
 	// special treatment for ungrouped data
 	if (count ($groupList) == 1 and isset ($groupList['other']))
 	{
-		printSelect ($groupList['other'], $select_name, $selected_id, $tabindex);
+		printSelect ($groupList['other'], $select_attrs, $selected_id);
 		return;
 	}
-	echo "<select name=${select_name}" . ($tabindex ? " tabindex=${tabindex}" : '') . '>';
+	if (!array_key_exists ($select_attrs, 'name'))
+		return;
+	echo '<select';
+	foreach ($select_attrs as $attr_name => $attr_value)
+		echo " ${attr_name}=${attr_value}";
+	echo '>';
 	foreach ($groupList as $groupname => $groupdata)
 	{
 		echo "<optgroup label='${groupname}'>";
 		foreach ($groupdata as $dict_key => $dict_value)
 			echo "<option value='${dict_key}'" . ($dict_key == $selected_id ? ' selected' : '') . ">${dict_value}</option>";
-		echo "</optgroup>\n";
+		echo '</optgroup>';
 	}
-	echo "</select>";
+	echo '</select>';
 }
 
 // used by renderGridForm() and renderRackPage()
@@ -1085,7 +1095,7 @@ function renderPortsForObject ($object_id)
 		printImageHREF ('add', 'add a port', TRUE);
 		echo "</td><td><input type=text size=8 name=port_name tabindex=100></td>\n";
 		echo "<td><input type=text size=24 name=port_label tabindex=101></td><td>";
-		printNiftySelect (getNewPortTypeOptions(), 'port_type_id', $prefs['selected'], 102);
+		printNiftySelect (getNewPortTypeOptions(), array ('name' => 'port_type_id', 'tabindex' => 102), $prefs['selected']);
 		echo "<td><input type=text name=port_l2address tabindex=103></td>\n";
 		echo "<td colspan=3>&nbsp;</td><td>";
 		printImageHREF ('add', 'add a port', TRUE, 104);
@@ -1115,7 +1125,7 @@ function renderPortsForObject ($object_id)
 			echo '<td>';
 			if ($port['iif_id'] != 1)
 				echo '<label>' . $port['iif_name'] . ' ';
-			printSelect (getExistingPortTypeOptions ($port['id']), 'port_type_id', $port['oif_id']);
+			printSelect (getExistingPortTypeOptions ($port['id']), array ('name' => 'port_type_id'), $port['oif_id']);
 			if ($port['iif_id'] != 1)
 				echo '</label>';
 			echo '</td>';
@@ -1187,7 +1197,7 @@ function renderPortsForObject ($object_id)
 	echo '<option value=ssv1>SSV:&lt;interface name&gt; &lt;MAC address&gt;</option>';
 	echo "</select>";
 	echo 'Default port type: ';
-	printNiftySelect (getNewPortTypeOptions(), 'port_type', $prefs['selected'], 202);
+	printNiftySelect (getNewPortTypeOptions(), array ('name' => 'port_type', 'tabindex' => 202), $prefs['selected']);
 	echo "<input type=submit value='Parse output' tabindex=204><br>\n";
 	echo "<textarea name=input cols=100 rows=50 tabindex=203></textarea><br>\n";
 	echo '</form>';
@@ -1206,7 +1216,7 @@ function renderIPv4ForObject ($object_id)
 		echo "<td class=tdleft><input type='text' size='10' name='bond_name' tabindex=100></td>\n";
 		echo "<td class=tdleft><input type=text name='ip' tabindex=101></td>\n";
 		echo "<td colspan=2>&nbsp;</td><td>";
-		printSelect ($aat, 'bond_type', NULL, 102);
+		printSelect ($aat, array ('name' => 'bond_type', 'tabindex' => 102));
 		echo "</td><td>&nbsp;</td><td>";
 		printImageHREF ('add', 'allocate', TRUE, 103);
 		echo "</td></tr></form>";
@@ -1268,7 +1278,7 @@ function renderIPv4ForObject ($object_id)
 			}
 		}
 		echo '<td>';
-		printSelect ($aat, 'bond_type', $alloc['type']);
+		printSelect ($aat, array ('name' => 'bond_type'), $alloc['type']);
 		echo "</td><td>";
 		$prefix = '';
 		if ($alloc['addrinfo']['reserved'] == 'yes')
@@ -2525,9 +2535,9 @@ function renderIPv4AddressAllocations ($dottedquad)
 		echo "<tr><td>";
 		printImageHREF ('add', 'allocate', TRUE);
 		echo "</td><td>";
-		printSelect (getNarrowObjectList ('IPV4OBJ_LISTSRC'), 'object_id', NULL, 100);
+		printSelect (getNarrowObjectList ('IPV4OBJ_LISTSRC'), array ('name' => 'object_id', 'tabindex' => 100));
 		echo "</td><td><input type=text tabindex=101 name=bond_name size=10></td><td>";
-		printSelect ($aat, 'bond_type', NULL, 102);
+		printSelect ($aat, array ('name' => 'bond_type', 'tabindex' => 102));
 		echo "</td><td>";
 		printImageHREF ('add', 'allocate', TRUE, 103);
 		echo "</td></form></tr>";
@@ -2556,7 +2566,7 @@ function renderIPv4AddressAllocations ($dottedquad)
 			echo "</a></td>";
 			echo "<td><a href='".makeHref(array('page'=>'object', 'object_id'=>$bond['object_id'], 'hl_ipv4_addr'=>$dottedquad))."'>${bond['object_name']}</td>";
 			echo "<td><input type='text' name='bond_name' value='${bond['name']}' size=10></td><td>";
-			printSelect ($aat, 'bond_type', $bond['type']);
+			printSelect ($aat, array ('name' => 'bond_type'), $bond['type']);
 			echo "</td><td>";
 			printImageHREF ('save', 'Save changes', TRUE);
 			echo "</td></form></tr>\n";
@@ -2575,7 +2585,7 @@ function renderNATv4ForObject ($object_id)
 		echo "<tr align='center'><td>";
 		printImageHREF ('add', 'Add new NAT rule', TRUE);
 		echo '</td><td>';
-		printSelect (array ('TCP' => 'TCP', 'UDP' => 'UDP'), 'proto');
+		printSelect (array ('TCP' => 'TCP', 'UDP' => 'UDP'), array ('name' => 'proto'));
 		echo "<select name='localip' tabindex=1>";
 
 		foreach ($alloclist as $dottedquad => $alloc)
@@ -2712,7 +2722,7 @@ function renderAddMultipleObjectsForm ()
 	{
 		echo '<tr><td>';
 		// Don't employ DEFAULT_OBJECT_TYPE to avoid creating ghost records for pre-selected empty rows.
-		printNiftySelect ($typelist, "${i}_object_type_id", 0, $tabindex);
+		printNiftySelect ($typelist, array ('name' => "${i}_object_type_id", 'tabindex' => $tabindex), 0);
 		echo '</td>';
 		echo "<td><input type=text size=30 name=${i}_object_name tabindex=${tabindex}></td>";
 		echo "<td><input type=text size=30 name=${i}_object_label tabindex=${tabindex}></td>";
@@ -2736,7 +2746,7 @@ function renderAddMultipleObjectsForm ()
 	echo "<table border=0 align=center><tr><th>names</th><th>type</th></tr>";
 	echo "<tr><td rowspan=3><textarea name=namelist cols=40 rows=25>\n";
 	echo "</textarea></td><td valign=top>";
-	printNiftySelect ($typelist, "global_type_id", getConfigVar ('DEFAULT_OBJECT_TYPE'));
+	printNiftySelect ($typelist, array ('name' => 'global_type_id'), getConfigVar ('DEFAULT_OBJECT_TYPE'));
 	echo "</td></tr>";
 	echo "<tr><th>Tags</th></tr>";
 	echo "<tr><td valign=top>";
@@ -3442,7 +3452,7 @@ function renderEditAttributesForm ()
 		printImageHREF ('create', 'Create attribute', TRUE);
 		echo "</td><td><input type=text tabindex=100 name=attr_name></td><td>";
 		global $attrtypes;
-		printSelect ($attrtypes, 'attr_type', NULL, 101);
+		printSelect ($attrtypes, array ('name' => 'attr_type', 'tabindex' => 101));
 		echo '</td><td>';
 		printImageHREF ('add', 'Create attribute', TRUE, 102);
 		echo '</td></tr></form>';
@@ -3492,7 +3502,7 @@ function renderEditAttrMapForm ()
 		echo "</select></td><td class=tdleft>";
 		printImageHREF ('add', '', TRUE);
 		echo ' ';
-		printNiftySelect (cookOptgroups (readChapter (CHAP_OBJTYPE, 'o')), 'objtype_id', NULL, 101);
+		printNiftySelect (cookOptgroups (readChapter (CHAP_OBJTYPE, 'o')), array ('name' => 'objtype_id', 'tabindex' => 101));
 		echo ' <select name=chapter_no tabindex=102><option value=0>-- dictionary chapter for [D] attributes --</option>';
 		foreach (getChapterList() as $chapter)
 			if ($chapter['sticky'] != 'yes')
@@ -4152,7 +4162,7 @@ function renderRSPoolServerForm ($pool_id)
 		'ipvs_2' => 'ipvsadm -l -n (address and port)',
 		'ipvs_3' => 'ipvsadm -l -n (address, port and weight)',
 	);
-	printSelect ($formats, 'format', 'ssv_1');
+	printSelect ($formats, array ('name' => 'format'), 'ssv_1');
 	echo "</td><td><input type=submit value=Parse></td></tr>\n";
 	echo "<tr><td colspan=3><textarea name=rawtext cols=100 rows=50></textarea></td></tr>\n";
 	echo "</table>\n";
@@ -4170,11 +4180,11 @@ function renderRSPoolLBForm ($pool_id)
 		echo "<table cellspacing=0 cellpadding=5 align=center>";
 		printOpFormIntro ('addLB');
 		echo "<tr valign=top><th class=tdright>Load balancer</th><td class=tdleft>";
-		printSelect (getNarrowObjectList ('IPV4LB_LISTSRC'), 'object_id', NULL, 1);
+		printSelect (getNarrowObjectList ('IPV4LB_LISTSRC'), array ('name' => 'object_id', 'tabindex' => 1));
 		echo '</td><td class=tdcenter valign=middle rowspan=2>';
 		printImageHREF ('ADD', 'Configure LB', TRUE, 5);
 		echo '</td></tr><tr><th class=tdright>Virtual service</th><td class=tdleft>';
-		printSelect (getIPv4VSOptions(), 'vs_id', NULL, 2);
+		printSelect (getIPv4VSOptions(), array ('name' => 'vs_id', 'tabindex' => 2));
 		echo "</td></tr>\n";
 		echo "<tr><th class=tdright>VS config</th><td colspan=2><textarea tabindex=3 name=vsconfig rows=10 cols=80></textarea></td></tr>";
 		echo "<tr><th class=tdright>RS config</th><td colspan=2><textarea tabindex=4 name=rsconfig rows=10 cols=80></textarea></td></tr>";
@@ -4227,11 +4237,11 @@ function renderVServiceLBForm ($vs_id)
 		echo '<table cellspacing=0 cellpadding=5 align=center>';
 		printOpFormIntro ('addLB');
 		echo '<tr valign=top><th class=tdright>Load balancer</th><td class=tdleft>';
-		printSelect (getNarrowObjectList ('IPV4LB_LISTSRC'), 'object_id', NULL, 101);
+		printSelect (getNarrowObjectList ('IPV4LB_LISTSRC'), array ('name' => 'object_id', 'tabindex' => 101));
 		echo '</td><td rowspan=2 class=tdcenter valign=middle>';
 		printImageHREF ('ADD', 'Configure LB', TRUE, 105);
 		echo '</td></tr><tr><th class=tdright>RS pool</th><td class=tdleft>';
-		printSelect (getIPv4RSPoolOptions(), 'pool_id', NULL, 102);
+		printSelect (getIPv4RSPoolOptions(), array ('name' => 'pool_id', 'tabindex' => 102));
 		echo '</td></tr>';
 		echo '<tr><th class=tdright>VS config</th><td colspan=2><textarea tabindex=103 name=vsconfig rows=10 cols=80></textarea></td></tr>';
 		echo '<tr><th class=tdright>RS config</th><td colspan=2><textarea tabindex=104 name=rsconfig rows=10 cols=80></textarea></td></tr>';
@@ -4370,7 +4380,7 @@ function renderVSListEditForm ()
 		if ($default_port == 0)
 			$default_port = '';
 		echo "<td><input type=text name=vport size=5 value='${default_port}' tabindex=102></td><td>";
-		printSelect ($protocols, 'proto', 'TCP');
+		printSelect ($protocols, array ('name' => 'proto'), 'TCP');
 		echo '</td><td><input type=text name=name tabindex=104></td><td>';
 		printImageHREF ('CREATE', 'create virtual service', TRUE, 105);
 		echo "</td><td rowspan=3>";
@@ -5035,12 +5045,12 @@ function renderObjectSLB ($object_id)
 		echo '<table cellspacing=0 cellpadding=5 align=center>';
 		printOpFormIntro ('addLB');
 		echo '<tr><th class=tdright>Virtual service</th><td class=tdleft>';
-		printSelect (getIPv4VSOptions(), 'vs_id', NULL, 101);
+		printSelect (getIPv4VSOptions(), array ('name' => 'vs_id', 'tabindex' => 101));
 		echo '</td><td class=tdcenter valign=middle rowspan=2>';
 		printImageHREF ('ADD', 'Configure LB', TRUE, 105);
 		echo '</td></tr>';
 		echo '</tr><th class=tdright>RS pool</th><td class=tdleft>';
-		printSelect (getIPv4RSPoolOptions(), 'pool_id', NULL, 102);
+		printSelect (getIPv4RSPoolOptions(), array ('name' => 'pool_id', 'tabindex' => 102));
 		echo "</td></tr>";
 		echo '<tr><th class=tdright>VS config</th><td colspan=2><textarea tabindex=103 name=vsconfig rows=10 cols=80></textarea></td></tr>';
 		echo '<tr><th class=tdright>RS config</th><td colspan=2><textarea tabindex=104 name=rsconfig rows=10 cols=80></textarea></td></tr>';
@@ -5107,7 +5117,7 @@ function renderEditVService ($vsid)
 	echo "<tr><th class=tdright>VIP:</th><td class=tdleft><input tabindex=1 type=text name=vip value='${vsinfo['vip']}'></td></tr>\n";
 	echo "<tr><th class=tdright>port:</th><td class=tdleft><input tabindex=2 type=text name=vport value='${vsinfo['vport']}'></td></tr>\n";
 	echo "<tr><th class=tdright>proto:</th><td class=tdleft>";
-	printSelect (array ('TCP' => 'TCP', 'UDP' => 'UDP'), 'proto', $vsinfo['proto']);
+	printSelect (array ('TCP' => 'TCP', 'UDP' => 'UDP'), array ('name' => 'proto'), $vsinfo['proto']);
 	echo "</td></tr>\n";
 	echo "<tr><th class=tdright>name:</th><td class=tdleft><input tabindex=4 type=text name=name value='${vsinfo['name']}'></td></tr>\n";
 	echo "<tr><th class=tdright>VS config:</th><td class=tdleft><textarea tabindex=5 name=vsconfig rows=20 cols=80>${vsinfo['vsconfig']}</textarea></td></tr>\n";
@@ -5508,7 +5518,7 @@ function renderFilesForEntity ($entity_id)
 		printOpFormIntro ('linkFile');
 		echo "<table border=0 cellspacing=0 cellpadding='5' align='center'>\n";
 		echo '<tr><td class=tdleft>';
-		printSelect ($files, 'file_id');
+		printSelect ($files, array ('name' => 'file_id'));
 		echo '</td><td class=tdleft>';
 		printImageHREF ('ATTACH', 'Link file', TRUE);
 		echo '</td></tr></table>';
@@ -6147,9 +6157,9 @@ function renderPortIFCompatEditor()
 		echo '<tr><th class=tdleft>';
 		printImageHREF ('add', 'add pair', TRUE);
 		echo '</th><th class=tdleft>';
-		printSelect (getPortIIFOptions(), 'iif_id');
+		printSelect (getPortIIFOptions(), array ('name' => 'iif_id'));
 		echo '</th><th class=tdleft>';
-		printSelect (readChapter (CHAP_PORTTYPE), 'oif_id');
+		printSelect (readChapter (CHAP_PORTTYPE), array ('name' => 'oif_id'));
 		echo '</th></tr></form>';
 	}
